oci-subtitle-translation/requirements.txt
oci-sdk[ai,object_storage]>=2.110.0
pyyaml>=6.0.1 